refactor(Mobile): Change structure mobile project
