--- mobile/QuantumCode/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Tue Apr 05 23:19:40 CST 2022
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.2-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
