+++ mobile/app/.gitignore
