+++ mobile/app/proguard-rules.pro
