+++ mobile/gradle.properties
