+++ mobile/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Apr 07 20:01:03 CST 2022
+distributionBase=GRADLE_USER_HOME
+distributionUrl=https\://services.gradle.org/distributions/gradle-7.2-bin.zip
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
